+++ src/model_signing/_cli.py
@@ -53,6 +53,14 @@ _read_signature_option = click.option(
     help="Location of the signature file to verify.",
 )
 
+# Decorator for the commonly used option for the custom trust configuration.
+_trust_config_option = click.option(
+        "--trust_config",
+        type=pathlib.Path,
+        metavar="TRUST_CONFIG_PATH",
+        help="The client trust configuration to use",
+)
+
 # Decorator for the commonly used option to ignore certain paths
 _ignore_paths_option = click.option(
     "--ignore-paths",
@@ -199,6 +207,7 @@ def _sign() -> None:
 @_ignore_git_paths_option
 @_write_signature_option
 @_sigstore_staging_option
+@_trust_config_option
 @click.option(
     "--use_ambient_credentials",
     type=bool,
@@ -222,6 +231,7 @@ def _sign_sigstore(
     use_ambient_credentials: bool,
     use_staging: bool,
     identity_token: Optional[str] = None,
+    trust_config: Optional[pathlib.Path] = None,
 ) -> None:
     """Sign using Sigstore (DEFAULT signing method).
 
@@ -238,12 +248,25 @@ def _sign_sigstore(
     Sigstore allows users to use a staging instance for test-only signatures.
     Passing the `--use_staging` flag would use that instance instead of the
     production one.
+
+    Additionally, you can specify a custom trust configuration JSON file using
+    the `--trust_config` flag. This allows you to fully customize the PKI 
+    (Private Key Infrastructure) used in the signing process. By providing a 
+    `--trust_config`, you can define your own transparency logs, certificate 
+    authorities, and other trust settings, enabling full control over the 
+    trust model, including which PKI to use for signature verification.
+
+    If `--trust_config` is not provided, the default Sigstore instance is 
+    used, which is pre-configured with Sigstore’s own trusted transparency logs 
+    and certificate authorities. This provides a ready-to-use default trust model 
+    for most use cases but may not be suitable for custom or highly regulated environments.
     """
     try:
         model_signing.signing.Config().use_sigstore_signer(
             use_ambient_credentials=use_ambient_credentials,
             use_staging=use_staging,
             identity_token=identity_token,
+            trust_config=trust_config,
         ).set_hashing_config(
             model_signing.hashing.Config().set_ignored_paths(
                 paths=list(ignore_paths) + [signature],
@@ -462,6 +485,11 @@ def _verify() -> None:
     signature is assumed to be generated via Sigstore (as if invoking `sigstore`
     subcommand).
 
+    To enable verification with custom PKI configurations, use the `--trust_config` option. 
+    This allows you to specify your own set of trusted public keys, transparency logs, and 
+    certificate authorities for verifying the signature. If not provided, the default Sigstore 
+    instance and its associated public keys, logs, and authorities are used. 
+
     Use each subcommand's `--help` option for details on each mode.
     """
 
@@ -472,6 +500,7 @@ def _verify() -> None:
 @_ignore_paths_option
 @_ignore_git_paths_option
 @_sigstore_staging_option
+@_trust_config_option
 @click.option(
     "--identity",
     type=str,
@@ -494,6 +523,7 @@ def _verify_sigstore(
     identity: str,
     identity_provider: str,
     use_staging: bool,
+    trust_config: Optional[pathlib.Path] = None,
 ) -> None:
     """Verify using Sigstore (DEFAULT verification method).
 
@@ -510,6 +540,7 @@ def _verify_sigstore(
             identity=identity,
             oidc_issuer=identity_provider,
             use_staging=use_staging,
+            trust_config=trust_config,
         ).set_hashing_config(
             model_signing.hashing.Config().set_ignored_paths(
                 paths=list(ignore_paths) + [signature],