+++ src/model_signing/signing.py
@@ -79,9 +79,10 @@ def sign(model_path: hashing.PathLike, signature_path: hashing.PathLike):
 class Config:
     """Configuration to use when signing models.
 
-    Currently we support signing with Sigstore (public instance and staging
-    instance), signing with private keys and signing with signing certificates.
-    Other signing modes can be added in the future.
+    Currently, we support signing with Sigstore (both the public instance and staging instance),
+    signing with private keys, signing with signing certificates, and signing with custom PKI configurations 
+    using the `--trust_config` option. This allows users to bring their own trust configuration to sign and verify
+    models. Other signing modes may be added in the future.
     """
 
     def __init__(self):
@@ -122,6 +123,7 @@ class Config:
         use_ambient_credentials: bool = False,
         use_staging: bool = False,
         identity_token: Optional[str] = None,
+        trust_config: Optional[pathlib.Path] = None,
     ) -> Self:
         """Configures the signing to be performed with Sigstore.
 
@@ -140,6 +142,11 @@ class Config:
               is supposed to be set to True only when testing. Default is False.
             identity_token: An explicit identity token to use when signing,
               taking precedence over any ambient credential or OAuth workflow.
+            trust_config: A path to a custom trust configuration. When provided, 
+              the signature verification process will rely on the supplied 
+              PKI and trust configurations, instead of the default Sigstore setup. 
+              If not specified, the default Sigstore configuration is used.
+
 
         Return:
             The new signing configuration.
@@ -149,6 +156,7 @@ class Config:
             use_ambient_credentials=use_ambient_credentials,
             use_staging=use_staging,
             identity_token=identity_token,
+            trust_config=trust_config,
         )
         return self